+++ app/vlstorage/main.go
@@ -13,6 +13,7 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/flagutil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fs"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/httpserver"
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/httputil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/promauth"
 	"github.com/VictoriaMetrics/metrics"
@@ -51,6 +52,8 @@ var (
 	minFreeDiskSpaceBytes = flagutil.NewBytes("storage.minFreeDiskSpaceBytes", 10e6, "The minimum free disk space at -storageDataPath after which "+
 		"the storage stops accepting new data")
 
+	logNewStreamsAuthKey = flagutil.NewPassword("logNewStreamsAuthKey", "authKey, which must be passed in query string to /internal/log_new_streams . It overrides -httpAuth.* . "+
+		"See https://docs.victoriametrics.com/victorialogs/#logging-new-streams")
 	forceMergeAuthKey = flagutil.NewPassword("forceMergeAuthKey", "authKey, which must be passed in query string to /internal/force_merge . It overrides -httpAuth.* . "+
 		"See https://docs.victoriametrics.com/victorialogs/#forced-merge")
 	forceFlushAuthKey = flagutil.NewPassword("forceFlushAuthKey", "authKey, which must be passed in query string to /internal/force_flush . It overrides -httpAuth.* . "+
@@ -235,6 +238,8 @@ func Stop() {
 func RequestHandler(w http.ResponseWriter, r *http.Request) bool {
 	path := r.URL.Path
 	switch path {
+	case "/internal/log_new_streams":
+		return processLogNewStreams(w, r)
 	case "/internal/force_merge":
 		return processForceMerge(w, r)
 	case "/internal/force_flush":
@@ -251,6 +256,29 @@ func RequestHandler(w http.ResponseWriter, r *http.Request) bool {
 		return processPartitionSnapshotList(w, r)
 	}
 	return false
+}
+
+func processLogNewStreams(w http.ResponseWriter, r *http.Request) bool {
+	if localStorage == nil {
+		// logging of new streams is available only at local storage
+		return false
+	}
+
+	if !httpserver.CheckAuthFlag(w, r, logNewStreamsAuthKey) {
+		return true
+	}
+
+	seconds, err := httputil.GetInt(r, "seconds")
+	if err != nil {
+		httpserver.Errorf(w, r, "cannot parse 'seconds' query arg: %s", err)
+		return true
+	}
+	if seconds <= 0 {
+		seconds = 10
+	}
+
+	localStorage.EnableLogNewStreams(seconds)
+	return true
 }
 
 func processForceMerge(w http.ResponseWriter, r *http.Request) bool {